crash-utility/arm64: introduce a dedicated field to record the mem layout changes
At present, we have the following important changes for arm64 memory
layout:

-1. redesigned ARM64 kernel virtual memory layout and associated KASLR
    support that was introduced in Linux 4.6. And NEW_VMEMMAP is used to
    flag it.
-2. memory layout flipped just right before introducing 52-bits kernel.
-3. introducing of vabits_actual and phyvirt_offset in kernel
-4. removing phyvirt_offset.

These changes have effects on PTOV()/VTOP() formula. So introducing a
dedicate field mmlayout_flags to record it.

Among above, 2 and 3 are introduced closely, and are not distinguished
in current implement. And this patch also keep this practice and use
vabits_actual as a hint to flag mem flipped.

Signed-off-by: Pingfan Liu <piliu@redhat.com>
Cc: HAGIO KAZUHITO <k-hagio-ab@nec.com>
Cc: Lianbo Jiang <lijiang@redhat.com>
Cc: Bhupesh Sharma <bhupesh.sharma@linaro.org>
To: crash-utility@redhat.com