Fix failure of gathering task table on linux-next (Linux 6.5 and later?)
(The following commit is now in linux-next.)

Kernel commit 757777eef55b ("pid: Replace struct pid 1-element array
with flex-array") changed pid.numbers[1] to pid.numbers[].  With this,
the size of struct pid does not contain the size of struct upid:

  (gdb) ptype /o struct pid
  /* offset    |  size */  type = struct pid {
  /*    0      |     4 */    refcount_t count;
  ...
  /*   96      |     0 */    struct upid numbers[];
      ^^^^          ^^^
                             /* total size (bytes):   96 */
                           }                         ^^^^

As a result, in refresh_xarray_task_table(), crash does not read the
data of pid.numbers[0].ns and cannot gather the task table correctly.

  $ crash vmlinux vmcore
  ...
  WARNING: active task ffff936992ad0000 on cpu 1 not found in PID hash
  ...
  crash> ps -S
    RU: 9
  crash>

Increase the size of reading struct pid by SIZE(upid) in this case.

Signed-off-by: Kazuhito Hagio <k-hagio-ab@nec.com>