crash add log dmesg PRINTK_CALLER id support
Submission to Project: crash
Component: dmesg
Files: kernel.c printk.c symbols.c help.c defs.h
Code level patch applied against: 8.0.4++ - latest code pulled from
        https://github.com/crash-utility/crash.git
crash Issue #164
Patch Version #2: per review from Hagio Kazuhito <k-hagio-ab@nec.com>
Tested with Kernel version and makedumpfile version:
Linux Kernel Testing: Linux catalina 6.6.6 #4 SMP PREEMPT_DYNAMIC
                        Tue Dec 12 23:11:30 PST 2023 x86_64 GNU/Linux
                      Linux 5.4.264 #9 SMP
                        Thu Dec 21 07:00:08 PST 2023
makedumpfile Testing: makedumpfile: version 1.7.4++
                      (released on 6 Nov 2023)
                      Issue 13 for makedumpfile: adds support for
                      demsg PRINTK_CALLER id field patch applied
dmesg Testing: util-linux 2.39.3++
                      Issue 2609 for sys-utils dmesg: adds support for
                      dmesg PRINTK_CALLER id field to standard
                      dmesg kmsg interface patch applied

Add support so that dmesg entries include the optional Linux Kernel
debug CONFIG option PRINTK_CALLER which adds an optional dmesg field
that contains the Thread Id or CPU Id that is issuing the printk to
add the message to the kernel ring buffer. If enabled, this CONFIG
option makes debugging simpler as dmesg entries for a specific
thread or CPU can be recognized.

The dmesg command supports printing the PRINTK_CALLER field. The
old syslog format (dmesg -S) and recently support was added for dmesg
using /dev/kmsg interface with util-linux Issue #2609 as we upstreamed
a commit that is under review.

We've upstreamed a patch for makedumpfile that adds support for
the PRINTK_CALLER id field so it will be available with the
commands:

        makedumpfile --dump-dmesg /proc/vmcore dmesgfile
        makedumpfile --dump-dmesg -x vmlinux /proc/vmcore dmesgfile

The additional field provided by PRINTK_CALLER is only present
if it was configured for the Linux kernel on the running system. The
PRINTK_CALLER is a debug option and not configured by default so the
dmesg output will only change for those kernels where the option was
configured when the kernel was built. For users who went to the
trouble to configure PRINTK_CALLER and have the extra field available
for debugging, having dmesg print the field is very helpful and so
will be makedumpfile and so it would be very useful to have crash
support for dump analysis.

Size of the PRINTK_CALLER field is determined by the maximum number
tasks that can be run on the system which is limited by the value of
/proc/sys/kernel/pid_max as pid values are from 0 to value - 1.
This value determines the number of id digits needed by the caller id.
The PRINTK_CALLER field is printed as T<id> for a Task Id or C<id>
for a CPU Id for a printk in CPU context. The values are left space
padded and enclosed in parentheses such as:
         [    T123]   or   [     C16]

Displaying the PRINTK_CALLER field in the log/dmesg record output:
-----------------------------------------------------------------

Given the layout of log/dmesg records printed by crash, for example:

crash> log -m
...
[    0.000000] <7>e820: remove [mem 0xff000000-0xffffffff] reserved
[    0.000000] <6>SMBIOS 3.4.0 present.
...
[    0.014179] <6>Secure boot disabled
[    0.014179] <6>RAMDISK: [mem 0x3cf4f000-0x437bbfff]
...
[  663.328848] <6>sysrq: Trigger a crash
[  663.328859] <0>Kernel panic - not syncing: sysrq triggered crash

Our patch adds the PRINTK_CALLER field after the timestamp if the
printk_caller log / dmesg option (-c) is selected:

crash> log -m -c
...
[    0.014179] [     T1] <6>Secure boot disabled
[    0.014179] [    T29] <6>RAMDISK: [mem 0x3cf4f000-0x437bbfff]
...

This is consistent placement with dmesg and makedumpfile.

To produce dmesg output with the PRINTK_CALLER id included, we add
a new log / dmesg command option: -c

The PRINTK_CALLER id field is printed only if the -c option is selected.
The description of the log -c option that is seen in the help is:

crash> log help

log
dump system message buffer
[-Ttdmasc]

...
...

-c  Display the caller id field that identifies either the thread id or
    the CPU id (if in CPU context) that called printk(), if available.
    Generally available on Linux 5.1 to 5.9 kernels configured with
    CONFIG_PRINTK_CALLER or Linux 5.10 and later kernels.

Also seen in the help file :

  Display the caller id that identifies the thread id of the task (begins
  with 'T') or the processor id (begins with 'C' for in CPU context) that
  called printk(), if available.

    crash> log -c
    ...
    [    0.014179] [     T1] Secure boot disabled
    [    0.014179] [    T29] RAMDISK: [mem 0x3cf4f000-0x437bbfff]
    [    0.198789] [     C0] DMAR: DRHD: handling fault status reg 3
    ...

Signed-off-by: Ivan Delalande <colona@arista.com>
Signed-off-by: Edward Chron <echron@arista.com>